requirements.txt
django==5.*
djangorestframework
django-debug-toolbar
